feat: update 0.1.x entry points for standalone binary
